@@ -4,8 +4,10 @@
     "main": "index.js",
     "license": "MIT",
     "scripts": {
-        "start": "NODE_ENV=development webpack-dev-server --open --content-base dist --env=dev --progress",
-        "build": "rimraf ./dist && webpack -p",
+        "prestart": "tcm src",
+        "start": "ODE_ENV=development webpack-dev-server --open --content-base dist --env=dev --progress",
+        "prebuild": "tcm src && rimraf ./dist",
+        "build": "webpack -p",
         "prebuild:start": "yarn build",
         "build:start": "serve -s -o -C -n dist",
         "lint": "tslint --project ./tsconfig.json --config ./tslint.json",
@@ -17,60 +19,61 @@
     "devDependencies": {
         "@types/enzyme": "3.1.3",
         "@types/enzyme-to-json": "^1.5.0",
-        "@types/extract-text-webpack-plugin": "^2.1.0",
+        "@types/extract-text-webpack-plugin": "^3.0.0",
         "@types/html-webpack-plugin": "^2.28.0",
         "@types/jest": "21.1.5",
-        "@types/node": "^8.0.28",
+        "@types/node": "^8.0.47",
         "@types/optimize-css-assets-webpack-plugin": "1.3.1",
-        "@types/react": "16.0.19",
+        "@types/react": "16.0.20",
         "@types/react-dom": "16.0.2",
         "@types/react-hot-loader": "^3.0.5",
         "@types/react-test-renderer": "^16.0.0",
-        "@types/webpack": "^3.0.10",
-        "@types/webpack-env": "^1.13.1",
-        "awesome-typescript-loader": "^3.2.3",
+        "@types/webpack": "^3.0.14",
+        "@types/webpack-env": "^1.13.2",
+        "awesome-typescript-loader": "^3.3.0",
         "babel-core": "6.26.0",
         "babel-loader": "7.1.2",
         "babel-polyfill": "6.26.0",
         "babel-preset-es2015": "6.24.1",
         "babel-preset-react": "6.24.1",
         "babel-preset-stage-0": "6.24.1",
-        "css-loader": "^0.28.7",
-        "enzyme": "^3.0.0",
-        "enzyme-adapter-react-16": "^1.0.2",
+        "css-loader": "0.28.7",
+        "enzyme": "^3.1.0",
+        "enzyme-adapter-react-16": "^1.0.3",
         "enzyme-to-json": "3.2.1",
-        "extract-text-webpack-plugin": "^3.0.0",
+        "extract-text-webpack-plugin": "^3.0.2",
         "html-webpack-plugin": "^2.30.1",
         "husky": "^0.14.3",
         "identity-obj-proxy": "^3.0.0",
-        "jest": "^21.1.0",
+        "jest": "^21.2.1",
         "json-d-ts": "^1.0.1",
         "json-loader": "0.5.7",
-        "node-sass": "^4.5.3",
+        "node-sass": "^4.6.0",
         "optimize-css-assets-webpack-plugin": "3.2.0",
         "prop-types": "15.6.0",
         "raf": "^3.4.0",
-        "react-hot-loader": "^3.0.0-beta.7",
-        "react-test-renderer": "16",
-        "resolve-url-loader": "^2.1.0",
-        "rimraf": "^2.6.1",
+        "react-hot-loader": "3.1.1",
+        "react-test-renderer": "16.0.0",
+        "resolve-url-loader": "^2.2.0",
+        "rimraf": "^2.6.2",
         "sass-loader": "^6.0.6",
-        "serve": "6.3.1",
-        "source-map-loader": "^0.2.1",
-        "style-loader": "^0.18.2",
+        "serve": "6.4.1",
+        "source-map-loader": "^0.2.3",
+        "style-loader": "^0.19.0",
         "stylelint": "8.2.0",
         "stylelint-config-sass-guidelines": "3.1.0",
         "ts-jest": "21.1.4",
         "ts-node": "3.3.0",
-        "tslint": "^5.7.0",
+        "tslint": "^5.8.0",
         "tslint-react": "3.2.0",
-        "tsutils": "^2.6.1",
+        "tsutils": "^2.12.2",
+        "typed-css-modules": "0.3.1",
         "typescript": "2.6.1",
         "typescript-babel-jest": "1.0.5",
-        "typings-for-css-modules-loader": "^1.7.0",
-        "webpack": "^3.5.6",
-        "webpack-dev-server": "^2.7.1",
-        "webpack-hot-middleware": "^2.19.1"
+        "typings-for-css-modules-loader": "1.7.0",
+        "webpack": "^3.8.1",
+        "webpack-dev-server": "^2.9.4",
+        "webpack-hot-middleware": "^2.20.0"
     },
     "dependencies": {
         "react": "^16.0.0",
