@@ -1,9 +1,9 @@
+/* tslint:disable */
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 const loader = require('react-hot-loader');
 import App from './components/App';
 
-/* tslint:disable */
 ReactDOM.render(
     <loader.AppContainer>
         <App/>
